adjust file structure
