+++ be/src/util/CMakeLists.txt
@@ -44,6 +44,7 @@ set(UTIL_FILES
   hash_util.hpp
   json_util.cpp
   json.cpp
+  json_converter.cpp
   starrocks_metrics.cpp
   mem_info.cpp
   metrics.cpp
@@ -113,8 +114,8 @@ set(UTIL_FILES
 # simdjson Runtime Implement Dispatch: https://github.com/simdjson/simdjson/blob/master/doc/implementation-selection.md#runtime-cpu-detection
 # Unset architecture-specific flags to avoid breaking implement runtime dispatch.
 if ("${CMAKE_BUILD_TARGET_ARCH}" STREQUAL "x86" OR "${CMAKE_BUILD_TARGET_ARCH}" STREQUAL "x86_64")
-    set_source_files_properties(json.cpp PROPERTIES COMPILE_FLAGS -mno-avx)
-    set_source_files_properties(json.cpp PROPERTIES COMPILE_FLAGS -mno-avx2)
+    set_source_files_properties(json.cpp PROPERTIES COMPILE_FLAGS "-mno-avx -mno-avx2")
+    set_source_files_properties(json_converter.cpp PROPERTIES COMPILE_FLAGS "-mno-avx -mno-avx2")
 endif()
